pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "migration-intelligence-agent"
version = "0.1.0"
description = "Multi-step Google ADK pipeline for migration intelligence"
authors = [{name = "Migration Intelligence Team"}]
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "flask>=3.0.0",
    "httpx>=0.27.0",
    "pandas>=2.2.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.4.0",
    "statsmodels>=0.14.0",
    "duckdb>=1.0.0",
    "gunicorn>=23.0.0",
    "pyarrow>=15.0.0",
    "plotly>=5.18.0",
    "python-dotenv>=1.0.0",
    "geopy>=2.4.0",
    "pydantic>=2.6.0",
]

[project.scripts]
migration-intel = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["agents", "analysis", "models", "tools"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
